moved org.mozilla.mozstumbler.core package to org.mozilla.mozstumbler.service.core so patches to fennec are easier
