+++ android/src/main/java/org/mozilla/mozstumbler/client/Updater.java
@@ -15,7 +15,7 @@ import android.os.AsyncTask;
 import android.util.Log;
 
 import org.mozilla.mozstumbler.R;
-import org.mozilla.mozstumbler.core.http.IHttpUtil;
+import org.mozilla.mozstumbler.service.core.http.IHttpUtil;
 import org.mozilla.mozstumbler.service.AppGlobals;
 import org.mozilla.mozstumbler.service.utils.NetworkInfo;
