+++ android/src/main/java/org/mozilla/mozstumbler/client/mapview/MLSLocationGetter.java
@@ -11,9 +11,9 @@ import android.util.Log;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.mozilla.mozstumbler.service.utils.LocationAdapter;
-import org.mozilla.mozstumbler.core.http.ILocationService;
-import org.mozilla.mozstumbler.core.http.IResponse;
-import org.mozilla.mozstumbler.core.http.MLS;
+import org.mozilla.mozstumbler.service.core.http.ILocationService;
+import org.mozilla.mozstumbler.service.core.http.IResponse;
+import org.mozilla.mozstumbler.service.core.http.MLS;
 import org.mozilla.mozstumbler.service.AppGlobals;
 
 import java.util.concurrent.atomic.AtomicInteger;
