+++ android/src/main/java/org/mozilla/mozstumbler/client/mapview/MapActivity.java
@@ -21,8 +21,7 @@ import android.view.ViewGroup;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
-import java.net.URL;
-import java.util.Scanner;
+
 import java.util.Timer;
 import java.util.TimerTask;
 import org.json.JSONException;
@@ -32,8 +31,8 @@ import org.mozilla.mozstumbler.R;
 import org.mozilla.mozstumbler.client.ClientPrefs;
 import org.mozilla.mozstumbler.client.ClientStumblerService;
 import org.mozilla.mozstumbler.client.MainApp;
-import org.mozilla.mozstumbler.core.http.HttpUtil;
-import org.mozilla.mozstumbler.core.http.IHttpUtil;
+import org.mozilla.mozstumbler.service.core.http.HttpUtil;
+import org.mozilla.mozstumbler.service.core.http.IHttpUtil;
 import org.mozilla.mozstumbler.client.ObservedLocationsReceiver;
 import org.mozilla.mozstumbler.client.navdrawer.MetricsView;
 import org.mozilla.mozstumbler.service.AppGlobals;
