+++ android/src/main/java/org/mozilla/mozstumbler/service/core/http/IHttpUtil.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.mozstumbler.core.http;
+package org.mozilla.mozstumbler.service.core.http;
 
 import java.io.File;
 import java.io.IOException;
